insert test

+++ Week_07/week7-homework/src/main/java/com/example/week7homework/entity/Goods.java
@@ -14,7 +14,7 @@ public class Goods {
 
     private Integer goodsPrice;
 
-    private Byte pubStatus;
+    private Integer pubStatus;
 
     private Date createTime;
