+++ Week_07/week7-homework/src/main/java/com/example/week7homework/entity/Order.java
@@ -14,7 +14,7 @@ public class Order {
 
     private Integer payAmount;
 
-    private Byte payStatus;
+    private Integer payStatus;
 
     private Date payTime;
